@@ -21,9 +21,6 @@ use Symfony\Component\Form\FormEvent;
  */
 class TranslationsFormsListener extends KoffI18nListener
 {
-    /**
-     * @param FormEvent $event
-     */
     public function preSetData(FormEvent $event)
     {
         $form = $event->getForm();
@@ -34,7 +31,7 @@ class TranslationsFormsListener extends KoffI18nListener
                 $locale,
                 $formOptions['form_type'],
                 $formOptions['form_options'] + [
-                    'required' => in_array($locale, $formOptions['required_locales'], true),
+                    'required' => \in_array($locale, $formOptions['required_locales'], true),
                 ]
             );
         }
