@@ -13,26 +13,16 @@ class AutoFormType extends AbstractType
     /** @var autoFormListener */
     private $autoFormListener;
 
-    /**
-     * @param AutoFormListener $autoFormListener
-     */
     public function __construct(AutoFormListener $autoFormListener)
     {
         $this->autoFormListener = $autoFormListener;
     }
 
-    /**
-     * @param FormBuilderInterface $builder
-     * @param array                $options
-     */
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
         $builder->addEventSubscriber($this->autoFormListener);
     }
 
-    /**
-     * @param OptionsResolver $resolver
-     */
     public function configureOptions(OptionsResolver $resolver)
     {
         $resolver->setDefaults(['fields' => [], 'excluded_fields' => []]);
@@ -41,7 +31,7 @@ class AutoFormType extends AbstractType
             'data_class',
             function (Options $options, $value) {
                 if (empty($value)) {
-                    throw new \RuntimeException(sprintf('Missing "data_class" option of "AutoFormType".'));
+                    throw new \RuntimeException('Missing "data_class" option of "AutoFormType".');
                 }
 
                 return $value;
